# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /catalog/releases/current:
    get:
      operationId: getCurrentCatalogRelease
      tags: [catalog]
      summary: Get the current approved public catalog release
      description: Returns approved public product metadata and local asset references. Pricing is never included.
      responses:
        "200":
          description: Current catalog release
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicCatalogRelease"
        "503":
          description: No valid approved catalog release is available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pricebooks/current:
    get:
      operationId: getCurrentPricebook
      tags: [pricebook]
      summary: Get the current protected wholesale pricebook
      description: Requires established server-side dealer or Kessick authorization. This endpoint fails closed when no authorization integration exists.
      responses:
        "200":
          description: Authorized pricebook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProtectedPricebook"
        "403":
          description: Wholesale pricebook access is unavailable or forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: No approved pricebook is available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    KessickOAuth:
      type: oauth2
      description: Authorization Code with PKCE. Users authenticate and consent in Kessick; access tokens last 15 minutes and refresh tokens are revocable.
      flows:
        authorizationCode:
          authorizationUrl: /api/oauth/authorize
          tokenUrl: /api/oauth/token
          scopes:
            catalog:read: Read approved public catalog facts
            project:read: Read only projects authorized for the user
            concierge:ask: Submit credit-acknowledged grounded requests
            action:confirm: Confirm a reviewed action
            mcp:use: Use the read-first MCP adapter
    ConciergeBearer:
      type: http
      scheme: bearer
      bearerFormat: Kessick connection token
      description: Per-user, scoped, expiring and revocable token issued only after the user authenticates with Clerk in Kessick.
...
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required: [error]
...
    CatalogAsset:
      type: object
      additionalProperties: false
      properties:
        originalUrl:
          type: string
        thumbnailUrl:
          type: string
        sourceUrl:
          type: string
        sha256:
          type: string
          pattern: "^[a-f0-9]{64}$"
      required: [originalUrl, thumbnailUrl, sourceUrl, sha256]
    PublicCatalogProduct:
      type: object
      additionalProperties: false
      properties:
        id: { type: string, minLength: 1 }
        sku: { type: string, minLength: 1 }
        catalogRef: { type: ["string", "null"] }
        name: { type: string, minLength: 1 }
        series: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        productType: { type: string, minLength: 1 }
        bottleCapacity: { type: ["number", "null"], minimum: 0 }
        widthIn: { type: number, exclusiveMinimum: 0 }
        heightIn: { type: number, exclusiveMinimum: 0 }
        depthIn: { type: number, exclusiveMinimum: 0 }
        dimensionsRaw: { type: string }
        unit:
          type: string
          enum: [inches]
        materials:
          type: array
          items: { type: string }
        finishes:
          type: array
          items: { type: string }
        hardwareFinishes:
          type: array
          items: { type: string }
        features:
          type: array
          items: { type: string }
        customizable: { type: boolean }
        fullyAssembled: { type: boolean }
        productUrl: { type: string }
        catalogPdf: { type: string }
        notes: { type: string }
        asset:
          anyOf:
            - $ref: "#/components/schemas/CatalogAsset"
            - type: "null"
      required:
        - id
        - sku
        - catalogRef
        - name
        - series
        - category
        - productType
        - bottleCapacity
        - widthIn
        - heightIn
        - depthIn
        - dimensionsRaw
        - unit
        - materials
        - finishes
        - hardwareFinishes
        - features
        - customizable
        - fullyAssembled
        - productUrl
        - catalogPdf
        - notes
        - asset
    CatalogProvenance:
      type: object
      additionalProperties: false
      properties:
        source: { type: string, minLength: 1 }
        sourceUrls:
          type: array
          items: { type: string }
        generatedAt:
          type: string
          format: date-time
        sourceVersion: { type: ["string", "null"] }
      required: [source, sourceUrls, generatedAt, sourceVersion]
    PublicCatalogRelease:
      type: object
      additionalProperties: false
      properties:
        releaseId: { type: string, pattern: "^sha256:[a-f0-9]{64}$" }
        releaseVersion: { type: ["string", "null"] }
        effectiveFrom: { type: ["string", "null"], format: date }
        effectiveTo: { type: ["string", "null"], format: date }
        brand: { type: string }
        website: { type: string }
        disclaimer: { type: string }
        provenance:
          $ref: "#/components/schemas/CatalogProvenance"
        products:
          type: array
          items:
            $ref: "#/components/schemas/PublicCatalogProduct"
      required:
        - releaseId
        - releaseVersion
        - effectiveFrom
        - effectiveTo
        - brand
        - website
        - disclaimer
        - provenance
        - products
    PricebookLine:
      type: object
      additionalProperties: false
      properties:
        sku: { type: string }
        currency: { type: string, pattern: "^[A-Z]{3}$" }
        wholesaleAmount:
          type: ["number", "null"]
          minimum: 0
        pricingStatus:
          type: string
          enum: [priced, tbd, unavailable]
      required: [sku, currency, wholesaleAmount, pricingStatus]
    ProtectedPricebook:
      type: object
      additionalProperties: false
      properties:
        version: { type: string }
        catalogReleaseId: { type: string }
        effectiveFrom: { type: string, format: date }
        effectiveTo: { type: ["string", "null"], format: date }
        provenance:
          $ref: "#/components/schemas/CatalogProvenance"
        lines:
          type: array
          items:
            $ref: "#/components/schemas/PricebookLine"
      required: [version, catalogReleaseId, effectiveFrom, effectiveTo, provenance, lines]
...